changed type variable and baudrate

+++ lib/Moti/Constants.h
@@ -10,20 +10,20 @@
 	const int DEFAULT_DELTA_ACCEL_THRESHOLD   = 200;
 	const int DEFAULT_HIGH_ACTIVITY_THRESHOLD = 80;
 
-	const int DEFAULT_LED_MAX_BRIGHTNESS   = 255;
-	const int DEFAULT_RED_MAX_BRIGHTNESS   = 255;
-	const int DEFAULT_GREEN_MAX_BRIGHTNESS = 255;
-	const int DEFAULT_BLUE_MAX_BRIGHTNESS  = 255;
+	const uint8_t DEFAULT_LED_MAX_BRIGHTNESS   = 255;
+	const uint8_t DEFAULT_RED_MAX_BRIGHTNESS   = 255;
+	const uint8_t DEFAULT_GREEN_MAX_BRIGHTNESS = 255;
+	const uint8_t DEFAULT_BLUE_MAX_BRIGHTNESS  = 255;
 
 	//	MOTOR PINS
-	const int leftMotorSpeedPin      = 5;
-	const int leftMotorDirectionPin  = 4;
-	const int rightMotorSpeedPin     = 6;
-	const int rightMotorDirectionPin = 7;
+	const uint8_t leftMotorSpeedPin      = 5;
+	const uint8_t leftMotorDirectionPin  = 4;
+	const uint8_t rightMotorSpeedPin     = 6;
+	const uint8_t rightMotorDirectionPin = 7;
 
 	//	LED PINS
-	const int RED_PIN   = 9;
-	const int GREEN_PIN = 10;
-	const int BLUE_PIN  = 11;
+	const uint8_t RED_PIN   = 9;
+	const uint8_t GREEN_PIN = 10;
+	const uint8_t BLUE_PIN  = 11;
 
 #endif
